moved functionality to the data preparation
